showreview/gateway/requirements.txt
psycopg2-binary
django-redis
django
requests
djangorestframework

